+++ broker/src/asapo_broker/server/post_op_image.go
@@ -2,13 +2,15 @@ package server
 
 import (
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
 	"net/http"
+	"strconv"
 )
 
 type ImageOp struct {
+	Id int
 	Op string
+	Params map[string]interface{} `json:",omitempty"`
 }
 func routeImageOp(w http.ResponseWriter, r *http.Request) {
 	body, err := ioutil.ReadAll(r.Body)
@@ -17,8 +19,14 @@ func routeImageOp(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	id, ok := extractRequestParametersID(r)
+	id_str, ok := extractRequestParametersID(r)
 	if !ok {
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	id, err := strconv.Atoi(id_str)
+	if err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
@@ -29,8 +37,7 @@ func routeImageOp(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-
-	fmt.Println(op)
-
-	processRequest(w, r, "ackimage", id, true)
+	op.Id = id
+	bOp,_ := json.Marshal(&op)
+	processRequest(w, r, op.Op, string(bOp), true)
 }