+++ broker/src/asapo_broker/server/server_nottested.go
@@ -53,8 +53,8 @@ func ReadConfig(fname string) (log.Level, error) {
 		return log.FatalLevel, errors.New("Server port not set")
 	}
 
-	if settings.CheckResendInterval<0  {
-		return log.FatalLevel, errors.New("Resend interval must be not negative")
+	if settings.CheckResendInterval==nil || *settings.CheckResendInterval<0  {
+		return log.FatalLevel, errors.New("Resend interval must be set and not negative")
 	}
 
 	if settings.PerformanceDbName == "" {